feat(User): 删除角色接口开发

+++ src/modules/user/user.controller.ts
@@ -1,13 +1,25 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  Query,
+  Res,
+  NotFoundException,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto, FindAllUserDto } from './dto';
 import { UpdateUserDto } from './dto';
 import { ApiBearerAuth, ApiOperation, ApiParam, ApiTags } from '@nestjs/swagger';
 import type { UserData } from './types';
 import { Roles, ReqUser } from '@/common/decorators';
-import { RoleEnum } from '@/helper/enums';
+import { ApiResponseCodeEnum, RoleEnum } from '@/helper/enums';
 import { User, Role } from '@/common/entities';
 import { ParseIntParamPipe } from '@/common/pipe';
+import { Response } from 'express';
 
 @ApiTags('User')
 @ApiBearerAuth()
@@ -49,7 +61,19 @@ export class UserController {
 
   @Delete(':id')
   @ApiOperation({ summary: '通过ID删除用户' })
-  async remove(@Param('id', new ParseIntParamPipe('id参数有误')) id: number) {
-    return this.usersService.remove(id);
+  async remove(
+    @Param('id', new ParseIntParamPipe('id参数有误')) id: number,
+    @Res({ passthrough: true }) res: Response,
+  ) {
+    const user = await this.usersService.findOneById(id).catch(() => false);
+    if (!user)
+      throw new NotFoundException({
+        code: ApiResponseCodeEnum.NOTFOUND_USER,
+        msg: '删除失败，未找到相关用户信息',
+      });
+    const delResult = await this.usersService.remove(id);
+    if (!delResult) res.resMsg = '删除用户失败!';
+    if (!delResult) res.success = false;
+    else return '删除用户成功';
   }
 }
